Bug 1265503 - Move data cycling tests into their own file
